REFACTOR: utils 폴더 구조 수정
- mocks 폴더 하위로 mockInfo 파일 이동
